Make the list sortable
A few things going on here:

* Assume that items from the server come back with a soldAt
and a scheduledAt property populated with times
* Anything in an App.List can be sorted by a property name,
(mercifully we only care about one direction). If the property
is a function, invoke it, otherwise sort by it naively.
* Because distance-from-center changes all the time, implement
`distance()` on App.Item. This required me to pull in the 
"geometry" lib from the google maps API (see the churn on index.us)
* @kpfefferle will appreciate my awesome buttons which lack
any style


Fixes #4